Add flow from db for monthly recaps

+++ shared/src/commonMain/kotlin/domain/entities/BalanceRecap.kt
@@ -1,7 +1,7 @@
 package domain.entities
 
 data class BalanceRecap(
-    val totalBalance: Int,
-    val monthlyIncome: Int,
-    val monthlyExpenses: Int
+    val totalBalance: Double,
+    val monthlyIncome: Double,
+    val monthlyExpenses: Double
 )
